+++ crates/router/src/connector/shift4.rs
@@ -93,6 +93,19 @@ impl ConnectorCommon for Shift4 {
 }
 
 impl api::Payment for Shift4 {}
+
+impl api::PaymentToken for Shift4 {}
+
+impl
+    ConnectorIntegration<
+        api::PaymentMethodToken,
+        types::PaymentMethodTokenizationData,
+        types::PaymentsResponseData,
+    > for Shift4
+{
+    // Not Implemented (R)
+}
+
 impl api::ConnectorAccessToken for Shift4 {}
 
 impl ConnectorIntegration<api::AccessTokenAuth, types::AccessTokenRequestData, types::AccessToken>