+++ crates/router/src/core/errors.rs
@@ -309,6 +309,8 @@ pub enum VaultError {
     MissingRequiredField { field_name: &'static str },
     #[error("The card vault returned an unexpected response: {0:?}")]
     UnexpectedResponseError(bytes::Bytes),
+    #[error("Failed to update in PMD table")]
+    UpdateInPaymentMethodDataTableFailed,
 }
 
 #[derive(Debug, thiserror::Error)]