+++ crates/router_env/src/build.rs
@@ -1,5 +1,0 @@
-include!("vergen.rs");
-
-fn main() {
-    generate_cargo_instructions();
-}