+++ migrations/2023-04-05-121040_alter_mca_change_country_to_enum/up.sql
@@ -1,0 +1,2 @@
+ALTER TABLE merchant_connector_account
+ALTER COLUMN business_country TYPE "CountryCode" USING business_country::"CountryCode";